+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+* Moved `test_extracted_archive_display` from the unit tests to a separate script.
+* Removed `test_extracted_archive_display`'s dependency on Tkinter.
+
 ## [0.3.2] - 2021-09-26
 ### Added
 * The module version number in `webarchive.__version__`.